refactor: modernize project Including typescript, functional React components and hooks, nanostore.
