+++ src/data/index.ts
@@ -1,0 +1,2 @@
+
+export const END_DATE = '2023-12-24T00:00:00';
